[refactoring] reconsiderate the folder structure
