+++ src/app/chain-game/component/game/game.component.scss
