+++ src/app/chain-game/component/game/history/history.component.html
