+++ src/app/chain-game/component/game/input/input.component.spec.ts
