+++ src/app/chain-game/component/game/now-word/now-word.component.html
