+++ src/app/chain-game/component/game/now-word/now-word.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { Chain } from 'src/app/model/Chain';
+import { Chain } from 'src/app/chain-game/model/Chain';
 
 @Component({
   selector: 'app-now-word',
