+++ src/app/chain-game/component/start/start.component.ts
