+++ src/app/component/site-frame/site-header/site-header.component.scss
